move stuff into baldr
